Move the new files to the new module
