+++ datacollection/build.gradle
@@ -33,7 +33,9 @@ android {
 }
 
 dependencies {
-
+    api ("ca.uhn.hapi.fhir:hapi-fhir-structures-r4:$hapiFhirStructuresR4Version") {
+        exclude module:'junit'
+    }
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlinVersion"
     implementation 'androidx.core:core-ktx:1.2.0'
     implementation 'androidx.appcompat:appcompat:1.2.0'
