+++ datacollection/src/main/java/com/google/android/fhir/datacollection/QuestionnaireViewModel.kt
@@ -1,4 +1,4 @@
-package com.google.fhirengine.ui
+package com.google.android.fhir.datacollection
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
@@ -18,34 +18,34 @@ class QuestionnaireViewModel(val questionnaire: Questionnaire) : ViewModel() {
     }
 
     private fun createQuestionnaireResponseItemComponent(
-            questionnaireItemComponent: Questionnaire.QuestionnaireItemComponent,
-            questionnaireResponseItemComponentList: MutableList<QuestionnaireResponse.QuestionnaireResponseItemComponent>) {
+        questionnaireItemComponent: Questionnaire.QuestionnaireItemComponent,
+        questionnaireResponseItemComponentList: MutableList<QuestionnaireResponse.QuestionnaireResponseItemComponent>) {
         when (questionnaireItemComponent.type) {
             Questionnaire.QuestionnaireItemType.BOOLEAN -> {
                 val questionnaireResponseItemComponent =
-                        QuestionnaireResponse.QuestionnaireResponseItemComponent(
-                                StringType(questionnaireItemComponent.linkId))
+                    QuestionnaireResponse.QuestionnaireResponseItemComponent(
+                        StringType(questionnaireItemComponent.linkId))
                 questionnaireResponseItemComponent.text = questionnaireItemComponent.text
                 questionnaireResponseItemComponentList.add(questionnaireResponseItemComponent)
                 map[questionnaireItemComponent.linkId] = questionnaireResponseItemComponent
             }
             Questionnaire.QuestionnaireItemType.STRING -> {
                 val questionnaireResponseItemComponent =
-                        QuestionnaireResponse.QuestionnaireResponseItemComponent(
-                                StringType(questionnaireItemComponent.linkId))
+                    QuestionnaireResponse.QuestionnaireResponseItemComponent(
+                        StringType(questionnaireItemComponent.linkId))
                 questionnaireResponseItemComponent.text = questionnaireItemComponent.text
                 questionnaireResponseItemComponentList.add(questionnaireResponseItemComponent)
                 map[questionnaireItemComponent.linkId] = questionnaireResponseItemComponent
             }
             Questionnaire.QuestionnaireItemType.GROUP -> {
                 val questionnaireResponseItemComponent =
-                        QuestionnaireResponse.QuestionnaireResponseItemComponent(
-                                StringType(questionnaireItemComponent.linkId))
+                    QuestionnaireResponse.QuestionnaireResponseItemComponent(
+                        StringType(questionnaireItemComponent.linkId))
                 questionnaireResponseItemComponent.text = questionnaireItemComponent.text
                 questionnaireResponseItemComponent.item = mutableListOf()
                 questionnaireItemComponent.item.forEach {
                     createQuestionnaireResponseItemComponent(it,
-                            questionnaireResponseItemComponent.item)
+                        questionnaireResponseItemComponent.item)
                 }
                 questionnaireResponseItemComponentList.add(questionnaireResponseItemComponent)
                 map[questionnaireItemComponent.linkId] = questionnaireResponseItemComponent
@@ -55,7 +55,7 @@ class QuestionnaireViewModel(val questionnaire: Questionnaire) : ViewModel() {
 }
 
 class QuestionnaireViewModelFactory(
-        private val questionnaire: Questionnaire
+    private val questionnaire: Questionnaire
 ) : ViewModelProvider.Factory {
     override fun <T : ViewModel?> create(modelClass: Class<T>): T {
         if (modelClass.isAssignableFrom(QuestionnaireViewModel::class.java)) {
